requirements.txt
fastapi
uvicorn
httpx
tortoise-orm
asyncpg
uuid6
python-dotenv
pydantic